Recreated gnugat/dotfiles project from scratch
gnugat/dotfiles is going to be completly rethought into a kinder project,
it starts by having a Contributor Code of Conduct.

See [Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/version/2/0/code_of_conduct/).

It is also using the Hippocratic License, which means that this project is Open
Source, but it is forbidden to use it in a way that promotes or profits from:

* violence, hate and division,
* environmental destruction,
* abuse of human rights,
* the destruction of people’s physical and mental health

See [Hippocratic License](https://firstdonoharm.dev/version/2/1/license.html).

It's also going to be organized a bit differently, with each tool having its own directory
containing their configuration and scripts to install it.